营销活动：完善 review 提到的问题

+++ yudao-module-mall/yudao-module-promotion-biz/src/main/java/cn/iocoder/yudao/module/promotion/service/article/ArticleCategoryServiceImpl.java
@@ -1,11 +1,13 @@
 package cn.iocoder.yudao.module.promotion.service.article;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.iocoder.yudao.framework.common.pojo.PageResult;
 import cn.iocoder.yudao.module.promotion.controller.admin.article.vo.category.ArticleCategoryCreateReqVO;
 import cn.iocoder.yudao.module.promotion.controller.admin.article.vo.category.ArticleCategoryPageReqVO;
 import cn.iocoder.yudao.module.promotion.controller.admin.article.vo.category.ArticleCategoryUpdateReqVO;
 import cn.iocoder.yudao.module.promotion.convert.article.ArticleCategoryConvert;
 import cn.iocoder.yudao.module.promotion.dal.dataobject.article.ArticleCategoryDO;
+import cn.iocoder.yudao.module.promotion.dal.dataobject.article.ArticleDO;
 import cn.iocoder.yudao.module.promotion.dal.mysql.article.ArticleCategoryMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
@@ -14,6 +16,7 @@ import javax.annotation.Resource;
 import java.util.List;
 
 import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
+import static cn.iocoder.yudao.module.promotion.enums.ErrorCodeConstants.ARTICLE_CATEGORY_DELETE_FAIL_HAVE_ARTICLES;
 import static cn.iocoder.yudao.module.promotion.enums.ErrorCodeConstants.ARTICLE_CATEGORY_NOT_EXISTS;
 
 /**
@@ -27,6 +30,8 @@ public class ArticleCategoryServiceImpl implements ArticleCategoryService {
 
     @Resource
     private ArticleCategoryMapper articleCategoryMapper;
+    @Resource
+    private ArticleService articleService;
 
     @Override
     public Long createArticleCategory(ArticleCategoryCreateReqVO createReqVO) {
@@ -50,7 +55,11 @@ public class ArticleCategoryServiceImpl implements ArticleCategoryService {
     public void deleteArticleCategory(Long id) {
         // 校验存在
         validateArticleCategoryExists(id);
-        // TODO @puhui999：需要校验下，是不是存在文章
+        // 校验是不是存在关联文章
+        List<ArticleDO> articleList = articleService.getArticleByCategoryId(id);
+        if (CollUtil.isNotEmpty(articleList)) {
+            throw exception(ARTICLE_CATEGORY_DELETE_FAIL_HAVE_ARTICLES);
+        }
 
         // 删除
         articleCategoryMapper.deleteById(id);
